+++ src/libs/apiConfig.js
@@ -2,6 +2,7 @@ const authApiBaseUrl = "https://jwt-auth-access-api.herokuapp.com";
 const authApiEndpointLiveness = "liveness";
 const authApiEndpointToken = "token";
 const baseApiUrl = "https://journal-notes-storage-api.herokuapp.com/json";
+const baseApiLiveness = "liveness";
 const baseApiConfig = { header: { "Content-Type": "application/json" } };
 const disableSave = false;
 const partialLoadCount = 25;
@@ -12,6 +13,7 @@ export {
   authApiEndpointToken,
   baseApiUrl,
   baseApiConfig,
+  baseApiLiveness,
   disableSave,
   partialLoadCount,
 };